fix: move density to cosmic_theme
