+++ Cargo.toml
@@ -84,11 +84,14 @@ derive_setters = "0.1.5"
 fraction = "0.15.3"
 image = { version = "0.25.1", optional = true }
 lazy_static = "1.4.0"
-libc = { version = "0.2.155", optional  = true }
+libc = { version = "0.2.155", optional = true }
 mime = { version = "0.3.17", optional = true }
 palette = "0.7.3"
 rfd = { version = "0.14.0", optional = true }
-rustix = { version = "0.38.34", features = ["pipe", "process"], optional = true }
+rustix = { version = "0.38.34", features = [
+    "pipe",
+    "process",
+], optional = true }
 serde = { version = "1.0.180", features = ["derive"] }
 slotmap = "1.0.6"
 smol = { version = "2.0.0", optional = true }
