+++ cosmic-theme/src/model/mod.rs
@@ -1,5 +1,6 @@
 pub use corner::*;
 pub use cosmic_palette::*;
+pub use density::*;
 pub use derivation::*;
 pub use mode::*;
 pub use spacing::*;
@@ -7,6 +8,7 @@ pub use theme::*;
 
 mod corner;
 mod cosmic_palette;
+mod density;
 mod derivation;
 mod mode;
 mod spacing;
